refactor(metrics): move yield-metric construction onto the collectors
SimplexMetricsCollector::into_yield_metric and
DuplexMetricsCollector::into_yield_metric (plus the moved
calculate_ideal_duplex_fraction_per_size) replace the separate-pass
commands' private functions, so the upcoming inline consensus path
calls the identical code. Duplex's real signature takes only
min_ab_reads/min_ba_reads - no third threshold parameter exists.

fgumi-metrics now depends on statrs directly for the moved binomial
CDF calculation. Pedantic lints (cast_precision_loss, float_cmp),
denied in this crate but only warned in the root fgumi crate, required
switching two ratio calculations to the existing frac() helper and
loosening a new test's float comparison to an epsilon check.
